Create README.md
The dataset is created to identify the locations of City-owned light poles and privately-owned utility poles that are approved for use by Mobile Telecommunications franchisees. Data is collected by DoITT following the reservation of each poles by a franchisaee and subsequent approval by the City. Interested parties may search the dataset to identify the location and certain other attributes of poles that may be used to support the expansion of mobile telecommunications services.